Reworked plugin completely
* Auto syncing is gone as it is now obsolete
* post_save listeners are gone
* Syncing between settings files now works correctly and completely
  autonomous

Reason for the problems in the earlier version was a misconception of the `key'
argument of the `add_on_change' and `clear_on_change' methods for settings. The
`key' argument refers to a unique identifier instead of an actual settings key.
All settings files now have correct callbacks which are added and removed
continuously to prevent loops and recursions.